Add feature model for folder structure. Create classe feature and move all related files
