+++ src/app/classes/class/[slug]/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import StudentsTable from "./StudentsTable";
+import StudentsTable from "../../../../features/classe/presentation/components/StudentsTable";
 
 export default function ClassPage({ params }: { params: { slug: string } }) {
   return <StudentsTable classId={params.slug} />;
