+++ src/core/auth/auth-store.ts
@@ -1,4 +1,3 @@
-import { get } from "lodash";
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
