+++ src/features/classe/domain/class-schema.ts
