+++ src/features/classe/infra/useGetClassesInfra.ts
@@ -1,8 +1,8 @@
 import { useQuery } from "convex/react";
-import { api } from "../../../convex/_generated/api";
-import { ClasseEntityDto } from "@/infrastructure/classe/classe-dto";
+import { api } from "../../../../convex/_generated/api";
+import { ClasseEntityDto } from "@/features/classe/infra/classe-dto";
 import { isRight } from "fp-ts/lib/Either";
-import ClassEntity from "@/domain/classe/class-entity";
+import ClassEntity from "@/features/classe/domain/class-entity";
 import { useEffect, useState } from "react";
 
 export default function useGetClassesInfra({ id }: { id: string }) {
