--- src/usecases/dependency-injection.usecases.ts
@@ -1,5 +1,0 @@
-// import ClassUsecases from "./class/class-usecases";
-// import { classRepositoryImpl } from "@/infrastructure/dependency-injection.infra";
-
-// export const classUsecases = new ClassUsecases(classRepositoryImpl);
-
